+++ public/app/features/alerting/unified/components/notification-policies/components/Modals.tsx
@@ -5,14 +5,6 @@ import { isFetchError } from '@grafana/runtime';
 import { Button, ConfirmModal, Modal, ModalProps, Space, Spinner, Stack, Text } from '@grafana/ui';
 
 import { RouteWithID } from '../../../../../../plugins/datasource/alertmanager/types';
-import {
-  trackNotificationPolicyCreateError,
-  trackNotificationPolicyCreated,
-  trackNotificationPolicyDeleteError,
-  trackNotificationPolicyDeleted,
-  trackNotificationPolicyReset,
-  trackNotificationPolicyResetError,
-} from '../../../Analytics';
 import { FormAmRoute } from '../../../types/amroutes';
 import { defaultGroupBy } from '../../../utils/amroutes';
 import { GRAFANA_RULES_SOURCE_NAME } from '../../../utils/datasource';
@@ -20,6 +12,14 @@ import { ROOT_ROUTE_NAME } from '../../../utils/k8s/constants';
 import { stringifyErrorLike } from '../../../utils/misc';
 import { AmRootRouteForm } from '../EditDefaultPolicyForm';
 import { NotificationPoliciesErrorAlert } from '../PolicyUpdateErrorAlert';
+import {
+  trackNotificationPolicyCreateError,
+  trackNotificationPolicyCreated,
+  trackNotificationPolicyDeleteError,
+  trackNotificationPolicyDeleted,
+  trackNotificationPolicyReset,
+  trackNotificationPolicyResetError,
+} from '../notificationPolicyAnalytics';
 
 export interface DeleteModalProps {
   isOpen: boolean;
